+++ fiddles/node/fiddle-0010-ZeroMqPubSub/client.js
@@ -1,0 +1,18 @@
+#!/usr/bin/env node --harmony
+
+'use strict';
+
+const zmq=require('zmq'),
+    subscriber = zmq.socket('sub');
+
+subscriber.subscribe('');
+subscriber.on('message', function(data) {
+    let message = JSON.parse(data),
+        date = new Date(message.timestamp);
+    console.log('File ' + message.file + ' changed at ' + date);
+});
+
+subscriber.connect("tcp://localhost:5432");
+
+
+